Use 'volatile' for VGA text buffer
Mark the actual memory representing VGA text buffer as volatile so that
Rust compiler doesn't peform any optimisations.

For e.g., Rust compiler doesn't know that we are actully writing to VGA
h/w; it might think we are writing to some memory and never reading that
data, and thus might optimise to remove the writes entirely!