Support GDT and separate stack for double-faults
Among other things, overflowing the kernel stack could result in a
triple-fault. This is because stack overflow corrupts the stack pointer
and thus double-fault handler cannot be launched correctly.
Triple-faults cannot be handled in SW and the processor resets itself.

To avoid this, we can use a separate stack for double-fault handler.
This can be configured via GDT and TSS.

This patch adds primitive support for GDT and sets-up a separate stack
(static buffer for now) for double-fault handler. Also, added a test.